+++ tests/particle/test_particle.py
@@ -359,7 +359,7 @@ checklist_describe = (
     [443, "Width = 0.0926 ± 0.0017 MeV"],  # J/psi
     # Test print-out of asymmetric width errors
     [4222, "Width = 1.89 + 0.09 - 0.18 MeV"],  # Sigma_c(2455)++
-    [23, "Width = 2495.2 ± 2.3 MeV"],  # H0
+    [23, "Width = 2495.5 ± 2.3 MeV"],  # H0
     # Test print-out of symmetric lifetime errors
     [5332, "Lifetime = 1.65e-03 + 1.8e-04 - 1.8e-04 ns"],  # Omega_b-
     [211, "Lifetime = 26.033 ± 0.005 ns"],  # pion
@@ -379,34 +379,30 @@ def test_describe(pid, description):
 
 
 def test_default_table_loading():
-    assert Particle.table_names() == ("particle2022.csv", "nuclei2022.csv")
+    assert Particle.table_names() == ("particle2023.csv", "nuclei2022.csv")
 
 
 def test_default_table_loading_bis():
     Particle.all()
     p = Particle.from_pdgid(211)
     assert p.table_loaded() is True
-    assert p.table_names() == ("particle2022.csv", "nuclei2022.csv")
+    assert p.table_names() == ("particle2023.csv", "nuclei2022.csv")
 
 
 def test_explicit_table_loading():
-    Particle.load_table(data.basepath / "particle2022.csv")
+    Particle.load_table(data.basepath / "particle2023.csv")
     assert Particle.table_loaded()
     assert len(Particle.table_names()) == 1
     assert Particle.all() is not None
 
 
 def test_all_particles_are_loaded():
-    Particle.load_table(data.basepath / "particle2018.csv")
-    assert len(Particle.all()) == 605
-    Particle.load_table(data.basepath / "particle2019.csv")
-    assert len(Particle.all()) == 610
-    Particle.load_table(data.basepath / "particle2020.csv")
-    assert len(Particle.all()) == 610
     Particle.load_table(data.basepath / "particle2021.csv")
     assert len(Particle.all()) == 616
     Particle.load_table(data.basepath / "particle2022.csv")
     assert len(Particle.all()) == 616
+    Particle.load_table(data.basepath / "particle2023.csv")
+    assert len(Particle.all()) == 622
 
     Particle.load_table(data.basepath / "nuclei2022.csv")
     assert len(Particle.all()) == 5880